requirements.txt
spidev
RPi.GPIO
numpy
websockets
opencv-contrib-python
flask